+++ rllib/agents/trainer_template.py
@@ -9,7 +9,7 @@ from ray.rllib.execution.metric_ops import StandardMetricsReporting
 from ray.rllib.policy import Policy
 from ray.rllib.utils import add_mixins
 from ray.rllib.utils.annotations import override, DeveloperAPI
-from ray.rllib.utils.types import TrainerConfigDict, ResultDict
+from ray.rllib.utils.typing import TrainerConfigDict, ResultDict
 
 logger = logging.getLogger(__name__)
