+++ rllib/evaluation/sample_collector.py
@@ -4,7 +4,7 @@ from typing import Dict, Optional
 
 from ray.rllib.evaluation.episode import MultiAgentEpisode
 from ray.rllib.models.modelv2 import ModelV2
-from ray.rllib.utils.types import AgentID, EpisodeID, PolicyID, \
+from ray.rllib.utils.typing import AgentID, EpisodeID, PolicyID, \
     TensorType
 
 logger = logging.getLogger(__name__)
