+++ rllib/models/torch/torch_action_dist.py
@@ -11,7 +11,7 @@ from ray.rllib.utils.numpy import SMALL_NUMBER, MIN_LOG_NN_OUTPUT, \
     MAX_LOG_NN_OUTPUT
 from ray.rllib.utils.spaces.space_utils import get_base_struct_from_space
 from ray.rllib.utils.torch_ops import atanh
-from ray.rllib.utils.types import TensorType, List
+from ray.rllib.utils.typing import TensorType, List
 
 torch, nn = try_import_torch()
