+++ rllib/offline/json_writer.py
@@ -16,7 +16,7 @@ from ray.rllib.offline.io_context import IOContext
 from ray.rllib.offline.output_writer import OutputWriter
 from ray.rllib.utils.annotations import override, PublicAPI
 from ray.rllib.utils.compression import pack, compression_supported
-from ray.rllib.utils.types import FileType, SampleBatchType
+from ray.rllib.utils.typing import FileType, SampleBatchType
 from typing import Any, List
 
 logger = logging.getLogger(__name__)
