+++ rllib/offline/mixed_input.py
@@ -4,7 +4,7 @@ from ray.rllib.offline.input_reader import InputReader
 from ray.rllib.offline.json_reader import JsonReader
 from ray.rllib.offline.io_context import IOContext
 from ray.rllib.utils.annotations import override, DeveloperAPI
-from ray.rllib.utils.types import SampleBatchType
+from ray.rllib.utils.typing import SampleBatchType
 from typing import Dict
