+++ rllib/offline/off_policy_estimator.py
@@ -5,7 +5,7 @@ from ray.rllib.policy.sample_batch import MultiAgentBatch, SampleBatch
 from ray.rllib.policy import Policy
 from ray.rllib.utils.annotations import DeveloperAPI
 from ray.rllib.offline.io_context import IOContext
-from ray.rllib.utils.types import TensorType, SampleBatchType
+from ray.rllib.utils.typing import TensorType, SampleBatchType
 from typing import List
 
 logger = logging.getLogger(__name__)
