+++ rllib/offline/wis_estimator.py
@@ -2,7 +2,7 @@ from ray.rllib.offline.off_policy_estimator import OffPolicyEstimator, \
     OffPolicyEstimate
 from ray.rllib.policy import Policy
 from ray.rllib.utils.annotations import override
-from ray.rllib.utils.types import SampleBatchType
+from ray.rllib.utils.typing import SampleBatchType
 
 
 class WeightedImportanceSamplingEstimator(OffPolicyEstimator):
