+++ rllib/utils/exploration/curiosity.py
@@ -27,7 +27,7 @@ from ray.rllib.models.torch.misc import SlimFC
 from ray.rllib.utils.exploration.exploration import Exploration
 from ray.rllib.utils.framework import try_import_torch, TensorType
 from ray.rllib.utils.from_config import from_config
-from ray.rllib.utils.types import SampleBatchType, TrainerConfigDict
+from ray.rllib.utils.typing import SampleBatchType, TrainerConfigDict
 
 torch, nn = try_import_torch()
