+++ rllib/utils/schedules/polynomial_schedule.py
@@ -3,7 +3,7 @@ from typing import Union
 from ray.rllib.utils.annotations import override
 from ray.rllib.utils.framework import try_import_tf, try_import_torch
 from ray.rllib.utils.schedules.schedule import Schedule
-from ray.rllib.utils.types import TensorType
+from ray.rllib.utils.typing import TensorType
 
 tf1, tf, tfv = try_import_tf()
 torch, _ = try_import_torch()
